Add card container component
